+++ cmps285HotelProject/HotelIntegratedComputerSystems/HotelIntegratedComputerSystems/Models/Employees/CustomersViewModel.cs
@@ -24,7 +24,7 @@ namespace HotelIntegratedComputerSystems.Models.Employees
         [DataType(DataType.PhoneNumber)]
         [Required(ErrorMessage = "Phone Number Not Provided")]
         [RegularExpression(@"^(\d{10})$", ErrorMessage = "Not A Valid Phone Number")]
-        public double Phone { get; set; }
+        public Int64 Phone { get; set; }
 
         [DisplayName("E-Mail")]
         [DataType(DataType.EmailAddress)]
